2 Human Players 1.0
First rough version of a working game. Both players are controlled by the user, and is just meant as a starting framework for creating later versions. The Card, Deck, Human classes and Player interface should be able to be used in the future for creating Human vs Computer games, and even Computer vs Computer games.