Adding Human Vs Npc with 2 Simple NPCs
Added a class for playing games against the computer. An abstract NPC class (implementing the Player interface) was created, currently not containing any additional methods. From that, two basic NPC classes were created, one always playing the lowest eligible card in hand and the other the highest. 
For the exisitng classes, some small housekeeping changes were made, such as remembering that toString() exists and putting that in, as well as some formatting changes for the TwoHumanGame interface.