Added NpcVsNpcGame as well as another Npc Class
NpcVsNpcGame allows you to play games against Npcs. Currently set to show the results of each round and who won. Plan to have a way to request a number of rounds be played, then print the results of each game in chart format. 
The new Npc class is RandomNpc, which plays completely randomly. Unlike the other simple Npcs, this one does use the Fox and Woodcutter abilities (selecting a card at random).