Added NpcVsNpcRoundSet and SimpleNpc
NpcVsNpcRoundSet makes 2 Npcs play a given amount of rounds together. Afterwards, the results are printed, detailing the scores and overall points awarded (ignoring Treasure/7 bonus). In the future, this should be modified in a way to only record results of certain hands, such as hands where the player starts with at least one Fox, or when the player starts with a hand summing to below/above a threshold.

SimpleNpc is my 1st attempt at a real Npc. Currently, it commits to either attempting to always win or always lose at the start of the round, based on the hand's sum and a threshold. The Npc understands the Witch, so that it will never be played and have the opposite effect of the computer's intent. The Swan and Monarch effects (which are more passive abilities that don't require much strategy) are simply ignored. The Fox, Woodcutter, and Treasure abilities are completely ignored. The next improvement on the schedule is to add dynamic strategy such that the computer can evaluate the game state (the round score)  and decide whether to change its strategy or not. So far, it has a round winrate over 60% against the RandomNpc (a round is won if the npc scored 6 points by finishing with 0-3 or 7-9 tricks).