Dynamic SimpleNpc, NpcVsNpcRoundSet Specifications
The SimpleNpc has been given a simple dynamic strategy. Currently works using a stored 2d array that uses the current round score to choose to aim high, aim low, or continue the current strategy. 
The NpcVsNpcRoundSet can now choose to only play rounds meeting certain requirements. The requirements that can be checked are the starting strategy of each NPC. This can be effective for gauging how well an NPC does with certain hands compared to others. In the future, more adjustable requirements can be added, such as having a certain amount of a specific card value.